pyproject.toml
[tool.poetry]    # содержит основную информацию о проекте
name = "podcast-bot"
version = "0.1.0"
description = ""
authors = ["Nikita Tokin <nikitatokin2006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]    # содержит описание всех зависимостей проекта
python = "^3.11"
aiogram = "3.*"
maturin = "1.*"
mypy = "1.5.*"

[tool.poetry.group.tests.dependencies]    # содержит описание всех зависимостей для тестов проекта
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
